requirements.txt
langchain~=0.3.26
langchain-core~=0.3.68

langchain-openai~=0.3.27
openai

langchain-anthropic~=0.3.17

langchain-google-genai~=2.1.6
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv~=1.1.1

numpy~=2.2.6
scikit-learn~=1.7.0
pip~=23.2.1
attrs~=25.3.0
distro~=1.9.0
protobuf~=5.29.5
pillow~=11.3.0
tornado~=6.5.1
Jinja2~=3.1.6
streamlit~=1.46.1
pydantic~=2.11.7
langchain_community
langchain-experimental~=0.3.4
langchain-community~=0.3.29
emoji~=2.14.1
PySocks~=1.7.1
cryptography~=45.0.7